Extract AddonHead component
