+++ src/amo/pages/Addon/index.js
@@ -4,13 +4,13 @@ import makeClassName from 'classnames';
 import * as React from 'react';
 import PropTypes from 'prop-types';
 import { compose } from 'redux';
-import Helmet from 'react-helmet';
 import { connect } from 'react-redux';
 
 import { setViewContext } from 'amo/actions/viewContext';
 import AddAddonToCollection from 'amo/components/AddAddonToCollection';
 import AddonBadges from 'amo/components/AddonBadges';
 import AddonCompatibilityError from 'amo/components/AddonCompatibilityError';
+import AddonHead from 'amo/components/AddonHead';
 import AddonMeta from 'amo/components/AddonMeta';
 import AddonMoreInfo from 'amo/components/AddonMoreInfo';
 import AddonRecommendations from 'amo/components/AddonRecommendations';
@@ -41,16 +41,15 @@ import {
   ADDON_TYPE_OPENSEARCH,
   ADDON_TYPE_STATIC_THEME,
   ADDON_TYPE_THEME,
-  CLIENT_APP_ANDROID,
   INCOMPATIBLE_NOT_FIREFOX,
   INSTALL_SOURCE_DETAIL_PAGE,
   UNKNOWN,
 } from 'core/constants';
 import { withInstallHelpers } from 'core/installAddon';
 import { isTheme, nl2br, sanitizeHTML, sanitizeUserHTML } from 'core/utils';
-import { getAddonJsonLinkedData, getErrorMessage } from 'core/utils/addons';
+import { getErrorMessage } from 'core/utils/addons';
 import { getClientCompatibility as _getClientCompatibility } from 'core/utils/compatibility';
-import { getAddonIconUrl, getPreviewImage } from 'core/imageUtils';
+import { getAddonIconUrl } from 'core/imageUtils';
 import translate from 'core/i18n/translate';
 import log from 'core/logger';
 import Button from 'ui/components/Button';
@@ -401,158 +400,6 @@ export class AddonBase extends React.Component {
     );
   }
 
-  getPageDescription() {
-    const { addon, i18n } = this.props;
-
-    return i18n.sprintf(
-      i18n.gettext('Download %(addonName)s for Firefox. %(summary)s'),
-      {
-        addonName: addon.name,
-        summary: addon.summary,
-      },
-    );
-  }
-
-  getPageTitle() {
-    const { addon, clientApp, i18n, lang } = this.props;
-
-    const i18nValues = {
-      addonName: addon.name,
-      locale: lang,
-    };
-
-    if (clientApp === CLIENT_APP_ANDROID) {
-      switch (addon.type) {
-        case ADDON_TYPE_DICT:
-          return i18n.sprintf(
-            // translators: please keep the fox emoji next to "Android".
-            i18n.gettext(`%(addonName)s – Get this Dictionary for 🦊 Android
-            (%(locale)s)`),
-            i18nValues,
-          );
-        case ADDON_TYPE_EXTENSION:
-          return i18n.sprintf(
-            // translators: please keep the fox emoji next to "Android".
-            i18n.gettext(`%(addonName)s – Get this Extension for 🦊 Android
-            (%(locale)s)`),
-            i18nValues,
-          );
-        case ADDON_TYPE_LANG:
-          return i18n.sprintf(
-            // translators: please keep the fox emoji next to "Android".
-            i18n.gettext(`%(addonName)s – Get this Language Pack for 🦊 Android
-            (%(locale)s)`),
-            i18nValues,
-          );
-        case ADDON_TYPE_STATIC_THEME:
-        case ADDON_TYPE_THEME:
-          return i18n.sprintf(
-            // translators: please keep the fox emoji next to "Android".
-            i18n.gettext(
-              `%(addonName)s – Get this Theme for 🦊 Android (%(locale)s)`,
-            ),
-            i18nValues,
-          );
-        case ADDON_TYPE_OPENSEARCH:
-          return i18n.sprintf(
-            // translators: please keep the fox emoji next to "Android".
-            i18n.gettext(`%(addonName)s – Get this Search Tool for 🦊 Android
-            (%(locale)s)`),
-            i18nValues,
-          );
-        default:
-          return i18n.sprintf(
-            // translators: please keep the fox emoji next to "Android".
-            i18n.gettext(`%(addonName)s – Get this Add-on for 🦊 Android
-            (%(locale)s)`),
-            i18nValues,
-          );
-      }
-    }
-
-    switch (addon.type) {
-      case ADDON_TYPE_DICT:
-        return i18n.sprintf(
-          // translators: please keep the fox emoji next to "Firefox".
-          i18n.gettext(`%(addonName)s – Get this Dictionary for 🦊 Firefox
-            (%(locale)s)`),
-          i18nValues,
-        );
-      case ADDON_TYPE_EXTENSION:
-        return i18n.sprintf(
-          // translators: please keep the fox emoji next to "Firefox".
-          i18n.gettext(`%(addonName)s – Get this Extension for 🦊 Firefox
-            (%(locale)s)`),
-          i18nValues,
-        );
-      case ADDON_TYPE_LANG:
-        return i18n.sprintf(
-          // translators: please keep the fox emoji next to "Firefox".
-          i18n.gettext(`%(addonName)s – Get this Language Pack for 🦊 Firefox
-            (%(locale)s)`),
-          i18nValues,
-        );
-      case ADDON_TYPE_STATIC_THEME:
-      case ADDON_TYPE_THEME:
-        return i18n.sprintf(
-          // translators: please keep the fox emoji next to "Firefox".
-          i18n.gettext(`%(addonName)s – Get this Theme for 🦊 Firefox
-            (%(locale)s)`),
-          i18nValues,
-        );
-      case ADDON_TYPE_OPENSEARCH:
-        return i18n.sprintf(
-          // translators: please keep the fox emoji next to "Firefox".
-          i18n.gettext(`%(addonName)s – Get this Search Tool for 🦊 Firefox
-            (%(locale)s)`),
-          i18nValues,
-        );
-      default:
-        return i18n.sprintf(
-          // translators: please keep the fox emoji next to "Firefox".
-          i18n.gettext(`%(addonName)s – Get this Add-on for 🦊 Firefox
-            (%(locale)s)`),
-          i18nValues,
-        );
-    }
-  }
-
-  renderMetaOpenGraph() {
-    const { addon, lang } = this.props;
-
-    const tags = [
-      <meta key="og:type" property="og:type" content="website" />,
-      <meta key="og:url" property="og:url" content={addon.url} />,
-      <meta key="og:title" property="og:title" content={this.getPageTitle()} />,
-      <meta
-        key="og:description"
-        property="og:description"
-        content={this.getPageDescription()}
-      />,
-      <meta key="og:locale" property="og:locale" content={lang} />,
-    ];
-
-    const image = addon.themeData
-      ? addon.themeData.previewURL
-      : getPreviewImage(addon);
-
-    if (image) {
-      tags.push(<meta key="og:image" property="og:image" content={image} />);
-    }
-
-    return tags;
-  }
-
-  renderJsonLinkedData() {
-    const { addon } = this.props;
-
-    return (
-      <script type="application/ld+json">
-        {JSON.stringify(getAddonJsonLinkedData({ addon }))}
-      </script>
-    );
-  }
-
   render() {
     const {
       addon,
@@ -645,19 +492,7 @@ export class AddonBase extends React.Component {
         })}
         data-site-identifier={addon ? addon.id : null}
       >
-        {addon && (
-          <Helmet titleTemplate={null}>
-            <title>{this.getPageTitle()}</title>
-            <link rel="canonical" href={addon.url} />
-            <meta name="description" content={this.getPageDescription()} />
-            <meta name="date" content={addon.created} />
-            {addon.last_updated && (
-              <meta name="last-modified" content={addon.last_updated} />
-            )}
-            {this.renderMetaOpenGraph()}
-            {this.renderJsonLinkedData()}
-          </Helmet>
-        )}
+        <AddonHead addon={addon} />
 
         {errorBanner}
